Fuzzy filename search for Baloo
Summary:
This is a proof of concept fuzzy search (only for filenames) based on bigrams. It's still not integrated into baloo or optimized, but I'm creating a diff to ask if I'm on the right track here or not. What's missing:

1. Integration into baloo as a search method
2. Ranking of fuzzy results
3. Journal of indexed files (to effeciently see if a file's been indexed yet)
4. Hook into `baloo_file` to gradually index files

Let me know what else I need or if you'd rather have a different fuzzy matching algorithm.

Reviewers: vhanda

Subscribers: kde-frameworks-devel, #baloo

Tags: #frameworks, #baloo

Differential Revision: https://phabricator.kde.org/D13747